scripts/requirements.txt
# Requirements for test scripts
# Install with: pip install -r scripts/requirements.txt

# Core dependencies
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Firebase dependencies
firebase-admin>=6.2.0

# Testing and utilities
asyncio-throttle>=1.0.2
colorama>=0.4.6

# Optional: For more advanced testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
